requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
openfga-sdk==0.9.5
aiosqlite==0.19.0
python-dotenv==1.0.0